Decided to generate new class Keymaker to handle key genreation, moved methods genreate shift and create offsets
