+++ test/cipher_key_test.rb
@@ -12,7 +12,7 @@ class CipherKeyTest <Minitest::Test
     assert_equal '040895', cipher_key.cipherdate
   end
 
-  def test_encrypt_can_create_cipherkey
+  def test_cipherkey_can_create_a_cipherkey
     cipher_key= CipherKey.new
     assert_equal 5, cipher_key.fill_in_cipherkey.length
 
@@ -32,33 +32,11 @@ class CipherKeyTest <Minitest::Test
     assert_equal '1025', cipher_key.last_four
   end
 
-  def test_encrypt_can_create_offsets
-    cipher_key= CipherKey.new('02715', '040895')
-    keys = [:A, :B, :C, :D]
-    expected = { A: 1,
-                 B: 0,
-                 C: 2,
-                 D: 5
-               }
-    assert_equal expected, cipher_key.create_offsets(keys)
-  end
-
   def test_can_pull_cipher_key_shift
     cipher_key= CipherKey.new('02715', '040895')
     keys = [:A, :B, :C, :D]
     key = :A
     assert_equal 2, cipher_key.cipher_shift(keys, key)
-  end
-
-  def test_encrypt_can_generate_keys
-    cipher_key= CipherKey.new('02715', '040895')
-    keys = [:A, :B, :C, :D]
-    expected = { A: 2,
-                 B: 27,
-                 C: 71,
-                 D: 15
-               }
-    assert_equal expected, cipher_key.generate_shift(keys)
   end
 
 end
